requirements.txt
asyncpg
duckdb
gcsfs
pyyaml
pyarrow
